Added bar attribute to ComboCoreChart
The bar attribute was missing from the ComboCoreChart so I went ahead 
and added it. Accepts an object with the key groupWidth which has a
value defined in the documentation as -

The width of a group of bars, specified in either of these formats:
 * Pixels (e.g. 50).
 * Percentage of the available width for each group (e.g. '20%'),
where '100%' means that groups have no space between them.